added support for style sheets and seperated the Build class methods into a Build class and a Render class
