+++ src/fjml/__init__.py
@@ -1,6 +1,6 @@
 from .parsers.builder import Build
 from .parsers.compiler import Compiler, ProgramLoader
-from .utils import Utilities, import_module, Validator, VALIDATOR_DTYPES, RegistryFileOperations
+from .utils import Utilities, Validator, VALIDATOR_DTYPES, RegistryFileOperations
 from .parsers.control_register import ControlRegistryOperations
 from .constant_controls.size_aware_control import SizeAwareControl
 from .constant_controls.custom_responsive_row import CustomResponsiveRow
