+++ src/fjml/parsers/compiler.py
@@ -9,7 +9,7 @@ from typing import (
     Final,
     NoReturn
 )
-
+import types
 import flet as ft
 
 
@@ -37,12 +37,13 @@ class Compiler:
         "control_awaitable", "program_name", "program", 
         "used_controls", "routes", "custom_controls",
         "control_bundles", "methods", "imports_path", "controls_registry",
-        "are_registries_joined"
+        "are_registries_joined", "style_sheet"
     )
     
     def __init__(
         self, program_name: str, code_input: dict[str, Any], 
-        custom_controls: list[dt.ControlJsonScheme] = [], imports_path: str = ""
+        custom_controls: list[dt.ControlRegisterInterface] = [], imports_path: str = "",
+        style_sheet: dt.JsonDict = {}
     ) -> NoReturn:
         """
         __init__ _summary_
@@ -51,8 +52,10 @@ class Compiler:
             program_name (str): _description_
         """
         self.used_controls: set[str] = set()
+        self.style_sheet: dt.StyleSheet = dt.StyleSheet(style_sheet)
+        self.update_used_controls(self.style_sheet.data)
         custom_controls = self.add_constant_controls(custom_controls)
-        self.custom_controls: dt.ControlRegistryJsonScheme = {}
+        self.custom_controls: dt.ControlRegistryJsonScheme
         if custom_controls:
             self.custom_controls = ControlRegistryOperations.generate_dict(
                 [dt.ControlRegistryModel(**control) for control in custom_controls],
@@ -61,16 +64,16 @@ class Compiler:
         self.are_registries_joined: bool = False
         self.imports_path: str = imports_path
         self.program_name: str = program_name
-        self.controls_registry: list[dt.ControlJsonScheme] = None
+        self.controls_registry: dt.ControlRegistryJsonScheme = dt.ControlRegistryJsonScheme()
         self.code: dt.JsonDict = code_input
         self.routes: set[str] = set()
         self.methods: dt.EventContainer
-        self.controls: dt.ControlMap = dt.ControlMap()
+        self.controls: dt.ControlMap = dt.TypeDict({}, (ft.Control, types.FunctionType, object))
         self.control_bundles: set[str] = set()
-        self.control_awaitable: dict[str, bool] = dict()
-        self.control_settings: dict[str, list[str]] = dict()
-        self.parsed_controls: dt.ParsedControls = dt.ParsedControls()
-        self.parsed_ui: dt.ParsedUserInterface = {}
+        self.control_awaitable: dict[str, bool] = dt.TypeDict({}, bool)
+        self.control_settings: dict[str, list[str]] = dt.TypeDict({}, list)
+        self.parsed_controls: dt.ParsedControls = dt.TypeDict({}, dt.ControlModel)
+        self.parsed_ui: dt.ParsedUserInterface = dt.TypeDict({}, dt.UserInterfaceViews)
     
     def validate_imports(self, file_name: str, data: dt.JsonDict) -> NoReturn:
         keys: list[str]
@@ -98,18 +101,20 @@ class Compiler:
             if key not in VALID_KEYS:
                 raise errors.InvalidMarkupContainerError("ui.json", key)
     
-    def add_constant_controls(self, custom_controls: list[dt.ControlJsonScheme]) -> list[dt.ControlRegistryDictPreview]:
+    def add_constant_controls(self, custom_controls: list[dt.ControlRegisterInterface]) -> list[dt.ControlRegisterInterface]:
         name: str
         for name in constant_controls.CONSTANT_CONTROLS:
             obj = getattr(constant_controls, name)
-            custom_controls.append({
-                "name":name,
-                "source":dt.ObjectSource(
-                    obj, obj.__module__
-                ),
-                "attr":name,
-                "is_awaitable":False
-            })
+            custom_controls.append(
+                dt.ControlRegisterInterface(
+                    name=name,
+                    source=dt.ObjectSource(
+                        obj, obj.__module__
+                    ),
+                    attr=name,
+                    is_awaitable=False
+                )
+            )
         
         return custom_controls
     
@@ -136,8 +141,6 @@ class Compiler:
                 continue
             
             raise ImportError(f"Control named, \"{name}\", is not registered")
-        
-        
     
     def __load_controls(self) -> NoReturn:
         if not self.controls_registry:
@@ -165,7 +168,7 @@ class Compiler:
         self.__parse_ui(self.code["UI"])
         
         return dt.CompiledModel(
-            self.parsed_controls, self.parsed_ui,
+            self.parsed_controls, self.style_sheet, self.parsed_ui,
             self.control_awaitable, self.controls,
             self.routes, self.control_bundles
         )
@@ -210,7 +213,7 @@ class Compiler:
         )
     
     def __parse_controls(self, control_data: list[dt.NamedControlDict]) -> dt.ParsedControls:
-        parsed_data: dt.ParsedControls = dt.ParsedControls()
+        parsed_data: dt.ParsedControls = dt.TypeDict({}, dt.ControlModel)
         var_name: str
         bundle_name: str
         bundles: list[str] = []
@@ -264,7 +267,7 @@ class Compiler:
         
         self.__load_controls()
     
-    def return_build(self, page: ft.Page, methods: dt.EventContainer, UserBuild: Optional[Type[Build]] = None) -> Build:
+    def return_build(self, page: ft.Page, methods: dt.EventContainer, user_build: Optional[Type[Build]] = None) -> Build:
         """
         return_build _summary_
         
@@ -277,10 +280,10 @@ class Compiler:
         compiled_program: dt.CompiledModel = self.compile()
         build: Build
         
-        if not UserBuild:
-            UserBuild = Build
+        if not user_build:
+            user_build = Build
         
-        build = UserBuild(compiled_program, page)
+        build = user_build(compiled_program, page)
         build.initialize()
         
         if methods:
@@ -293,14 +296,11 @@ class Compiler:
 
 def ProgramLoader(params: dt.LoaderParameters) -> Build:
     compiler: Compiler = Compiler(
-        params.program_name, 
-        params.ui_code,
-        params.custom_controls,
-        params.imports_path
+        params.program_name, params.ui_code,
+        params.custom_controls, params.imports_path,
+        params.style_sheet
     )
     
     return compiler.return_build(
-        params.page, 
-        params.methods,
-        params.UserBuild
+        params.page, params.methods, params.UserBuild
     )
