+++ src/fjml/utils.py
@@ -29,9 +29,6 @@ class ControlJsonScheme(TypedDict):
     name: str
     source: str
     attr: str
-    GET: list
-    POST: list
-    CALL: list
     awaitable: bool
     valid_settings: list[str]
 
@@ -47,7 +44,7 @@ import_module: Callable[[str, Optional[str]], ModuleType] = lru_cache(128)(impor
 class Utilities:
     
     @staticmethod
-    def get_object_args(func: Callable) -> list[str]:
+    def get_object_args(func: Callable[[...], Any]) -> list[str]:
         """
         get_object_args _summary_
         
@@ -251,20 +248,20 @@ class RegistryFileOperations:
     path: str = CONTROL_REGISTRY_PATH
 
     @classmethod
-    def load_file(self) -> ControlRegistryJsonScheme:
+    def load_file(cls) -> ControlRegistryJsonScheme:
         registry: io.TextIOWrapper
-        if not os.path.exists(self.path):
-            with open(self.path, 'w') as registry:
+        if not os.path.exists(cls.path):
+            with open(cls.path, 'w') as registry:
                 json.dump(EMPTY_REGISTRY_FILE, registry, indent=4)
-        with open(self.path, 'r') as registry:
+        with open(cls.path, 'r') as registry:
             return json.load(registry)
         raise RegistryFileNotFoundError()
 
     @classmethod
-    def save_file(self, file_data: JsonDict, dump_kwargs: dict[str, Any] = {}) -> NoReturn:
+    def save_file(cls, file_data: JsonDict, dump_kwargs: dict[str, Any] = {}) -> NoReturn:
         registry: io.TextIOWrapper
         if not dump_kwargs.get("indent", None):
             dump_kwargs["indent"] = 4
-        with open(self.path, 'w') as registry:
+        with open(cls.path, 'w') as registry:
             return json.dump(file_data, registry, **dump_kwargs)
         raise RegistryFileNotFoundError()
